+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Ability to add optional parameters when registering a daemon, for example *OVERRIDE_APP_HOST*. #269
 - Correct support of the Docker `HEALTHCHECK` instruction. #273
 - Support of pulling "custom" images for the selected compute device. #274
+- API for registering OCC commands.
 
 ### Fixed